12.22 / Authorization update, delete

+++ server/controller/tweet.js
@@ -29,16 +29,29 @@ export async function createTweet(req, res, next) {
 export async function updateTweet(req, res, next) {
   const id = req.params.id;
   const { text } = req.body;
-
-  const tweet = await tweetRepository.update(id, text);
+  const tweet = await tweetRepository.findById(id);
   if (!tweet) {
-    return res.status(404).json({ message: `Tweet id(${id}) not found` });
+    return res.sendStatus(404);
   }
-  res.status(200).json(tweet);
+  if (tweet.userId !== req.userId) {
+    return res.sendStatus(403);
+  }
+
+  const updated = await tweetRepository.update(id, text);
+
+  res.status(200).json(updated);
 }
 
 export async function deleteTweet(req, res, next) {
   const id = req.params.id;
+  const tweet = await tweetRepository.findById(id);
+  if (!tweet) {
+    return res.sendStatus(404);
+  }
+  if (tweet.userId !== req.userId) {
+    return res.sendStatus(403);
+  }
+
   await tweetRepository.remove(id);
   res.sendStatus(204);
 }
